cloudrun-services/requirements.txt
# FastAPI for web server
fastapi
uvicorn[standard]

# Google Cloud Storage for file uploads
google-cloud-storage

# Shap-E for 3D model generation
torch
filelock
requests
tqdm
ftfy
regex
numpy
Pillow
# Install the specific version of shap-e from OpenAI's GitHub
git+https://github.com/openai/shap-e.git

# Trimesh for 3D model conversion
trimesh

# For GPU support with PyTorch
# Ensure you have a compatible CUDA version in your Docker image
# This version is for CUDA 12.1, adjust if your base image has a different version
--extra-index-url https://download.pytorch.org/whl/cu121
torchvision
torchaudio